requirements.txt

pytest
pyyaml
